requirements.txt
streamlit
google-generativeai>=0.7.0
pandas
gspread
oauth2client
Pillow
google-api-python-client